requirements.txt
# PyTorch & torchvision
torch>=2.1.0
torchvision>=0.16.0

# Data handling
pandas>=2.1.0
numpy>=1.26.0

# Image processing
Pillow>=10.0.0
scikit-image>=0.20.0  # compatible with Python 3.11+

# Metrics
lpips>=0.1.4
pytorch-fid>=0.3.0  # latest compatible version with Python 3.11+

# Notebook & visualization
matplotlib>=3.8.0
tqdm>=4.66.0
jupyter>=1.0.0
ipykernel>=6.30.0  # compatible with Python 3.11
